Review resizing rendering and performances
When gallery loads, the page usually doesn't have a scroll but after adding items the scroll appear causing
a page resize. This cause two sizing computings : one before scroll appear, and another after.

To prevent that leak of performances the gallery is initialized with an height bigger than the rest of the page to
have a scroll before adding images to the DOM.

Additionnally, the gallery is now aware when resize starts and resize ends. When resize ends, the images are just hidden
to prevent user to see layout break.

When resize ends, the new container width is used to do a single computing for actually displayed images.