Add option for an individual focal point on images.
Gallery option: focalpoint: boolean
Image item property: focalpoint: string

If an image item has a focalpoint property, e.g. "20% 89%", this will be
applied as a background-position style on the thumbnail and the
background-size:cover is removed.

For items with focalpoint set, thumbnail size should be larger than
gallery option rowHeight.