Move to ESM
This change our own code to be (almost) entirely ESM code, as described
in https://github.com/sindresorhus/meta/discussions/15

And it also produce a second build as ESM to be consumed by Angular.

The browser build is presevered as-is, so we can still use the lib with
simple `<script>` tags.

Fixes #46